% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_outliers.R
\name{drop_outliers}
\alias{drop_outliers}
\title{Drop outliers}
\usage{
drop_outliers(data, var, prob, ...)
}
\arguments{
\item{data}{Input data}

\item{var}{Variable to use for calculating percentiles}

\item{prob}{Bottom & top percentile to drop. If 0.05, top and bottom 5 percentiles will be dropped.}

\item{...}{Further arguments to pass on \code{quantile}}
}
\value{
A tibble
}
\description{
Drop outliers based on certain percentiles
}
