% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census.R
\name{get_census}
\alias{get_census}
\title{Get Census Data}
\usage{
get_census(variable, CENSUS_API_KEY = Sys.getenv("CENSUS_API_KEY"))
}
\arguments{
\item{variable}{Census variable}

\item{CENSUS_API_KEY}{Census API Key}
}
\value{
A tibble of Census data.
}
\description{
Get Census data about population, income, etc.
}
