% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_char.R
\name{read_csv_char}
\alias{read_csv_char}
\title{Read all columns as character}
\usage{
read_csv_char(path)
}
\arguments{
\item{path}{file path}
}
\value{
A tibble
}
\description{
Read all columns using read_csv of readr as character.
}
